se modifican tipos de datos en productos

+++ src/main/java/Mintictest/ciclo3/dao/productosDAO.java
@@ -3,6 +3,6 @@ package Mintictest.ciclo3.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import Mintictest.ciclo3.model.productos;
 
-public interface productosDAO extends JpaRepository<productos, Integer>{
+public interface productosDAO extends JpaRepository<productos, Long>{
 
 }
